+++ sparse_strips/vello_hybrid/src/render/webgl/mod.rs
@@ -20,6 +20,9 @@
 only break in edge cases, and some of them are also only related to conversions from f64 to f32."
 )]
 
+#[cfg(feature = "probe")]
+pub(crate) mod probe;
+
 use crate::draw::ExternalTextureRun;
 use crate::render::common::IMAGE_PADDING;
 use crate::util::RangedSlice;
