+++ sparse_strips/vello_hybrid/src/render/webgl/probe.rs
